Group binding types implementations and converters in 'bindings' subpackage
