+++ azure/worker/dispatcher.py
@@ -12,11 +12,10 @@ import traceback
 
 import grpc
 
+from . import bindings
 from . import functions
 from . import loader
 from . import protos
-from . import type_impl
-from . import type_meta
 
 
 class DispatcherMeta(type):
@@ -240,16 +239,16 @@ class Dispatcher(metaclass=DispatcherMeta):
                     trigger_metadata = invoc_request.trigger_metadata
                 else:
                     trigger_metadata = None
-                params[pb.name] = type_meta.from_incoming_proto(
+                params[pb.name] = bindings.from_incoming_proto(
                     pb_type, pb.data, trigger_metadata)
 
             if fi.requires_context:
-                params['context'] = type_impl.Context(
+                params['context'] = bindings.Context(
                     fi.name, fi.directory, invocation_id)
 
             if fi.output_types:
                 for name in fi.output_types:
-                    params[name] = type_impl.Out()
+                    params[name] = bindings.Out()
 
             if fi.is_async:
                 call_result = await fi.func(**params)
@@ -267,7 +266,7 @@ class Dispatcher(metaclass=DispatcherMeta):
                         # Can "None" be marshaled into protos.TypedData?
                         continue
 
-                    rpc_val = type_meta.to_outgoing_proto(out_type, val)
+                    rpc_val = bindings.to_outgoing_proto(out_type, val)
                     assert rpc_val is not None
 
                     output_data.append(
@@ -277,7 +276,7 @@ class Dispatcher(metaclass=DispatcherMeta):
 
             return_value = None
             if fi.return_type is not None:
-                return_value = type_meta.to_outgoing_proto(
+                return_value = bindings.to_outgoing_proto(
                     fi.return_type, call_result)
 
             return protos.StreamingMessage(
